Stable API extensions
This introduces a new ZFS_IOC_LIBZFS_CORE ioctl and replaces the ioctl
calls in all but 1 function in libzfs_core.so and adds many more. It is
intended to be as stable kernel/userland interface that allows
basic userland functionality to continue working when an older userland
is used with newer kernel modules or vice versa by mistake.

At this time, it is a subset of the legacy functionality, but enough API
functions have been implemented that mismatched kernel/userland code
will not break the Gentoo's boot process (excluding zfs share and ZED,
which are excluded at this time) and ClusterHQ's flocker.

This sufficiently addresses the long standing /dev/zfs API instability
that it opens the possibility of stabilization of the ZFS packaging in
Gentoo on amd64 and other architectures, which makes this an improvement
in more ways than one.

Signed-off-by: Richard Yao <ryao@gentoo.org>